merge develop (#5)
* move some devDependencies to full-on dependencies

* remove npm-run-all module

It's really not that helpful since you can't effectively watch output of
both tasks (transpile and sync) in the single console window.

* add a LICENSE

* update package.json with licensing info

Can you tell I don't think of software licenses much?  Or is it not
thinking much of them? 🤔

* merge develop (#3)

* move some devDependencies to full-on dependencies

* remove npm-run-all module

It's really not that helpful since you can't effectively watch output of
both tasks (transpile and sync) in the single console window.

* add a LICENSE

* update package.json with licensing info

Can you tell I don't think of software licenses much?  Or is it not
thinking much of them? 🤔

* merge develop (#3) (#4)

* move some devDependencies to full-on dependencies

* remove npm-run-all module

It's really not that helpful since you can't effectively watch output of
both tasks (transpile and sync) in the single console window.

* add a LICENSE

* update package.json with licensing info

Can you tell I don't think of software licenses much?  Or is it not
thinking much of them? 🤔

* update package-lock.json

oops, missed yet another place for license information

* improve CommandFlag type

* change CommandFlags from interface to type

* reorder exports and make CommandFlag exportable

* add autocompletion helper function

* make AutoCompletionArgs completionKeys be optional

For cases when the only passed flags are booleans and do not require
values to be present since the flag is true if it exists and false
otherwise.

* add autocomplete to net mapper script

* refactor AutoComplete types for readability

sure, Jan!

* match Bitburner's casing

* export more AutoCompletion types

* add autocompletion to pwn script

* reverse the polarity of the loop check in wgh

This script is meant to loop by default, so use a negative condition
check on a passed boolean flag.  In my Bitburner world, you should
never need to pass a flag such as `--once false`

* add autocompletion to wgh script